+++ frontend/src/pages/publik/informasi/index.js
@@ -1,2 +1,3 @@
-export { default as KebijakanPrivasi } from './KebijakanPrivasi';
+export { default as Ihwal } from './Ihwal';
+export { default as Privasi } from './Privasi';
 export { default as Sumber } from './Sumber';